% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot}
\alias{biplot}
\title{Performs a plot crossing two variables, adapting to the kind of variable (qualitative or quantitative)}
\usage{
biplot(dataset, x, y, method = NULL, add_pval = TRUE)
}
\arguments{
\item{dataset}{the dataset containing the variables}

\item{x}{the name of first variable}

\item{y}{the name of second variable}

\item{add_pval}{defaults to TRUE, if TRUE, the p-value of the test is added to the title of the plot}

\item{in}{case X and Y are quantitative, the method of regression (defaults to NULL for non-linear regression, can be set to "lm" for a linear regression)}
}
\description{
Performs a plot crossing two variables, adapting to the kind of variable (qualitative or quantitative)
}
\examples{
# example code
data(icecream)
biplot(icecream,"creme_glacee_vanille","creme_glacee_chocolat")
biplot(icecream,"creme_glacee_chocolat","age")
biplot(icecream,"age","taille")
biplot(icecream,"age","taille",method=NULL)
}
