% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{bivar_type}
\alias{bivar_type}
\title{Assesses the types of variables (quantitative or qualitative) for a bivariate analysis}
\usage{
bivar_type(dataset, x, y)
}
\arguments{
\item{dataset}{the dataset containing the variables}

\item{x}{the name of first variable}

\item{y}{the name of second variable}
}
\value{
the type of bivariate analysis (either quanti-quanti, quali-quanti or quali-quali)
}
\description{
Assesses the types of variables (quantitative or qualitative) for a bivariate analysis
}
\examples{
# example code
data(icecream)
biplot(icecream,"creme_glacee_vanille","creme_glacee_chocolat")
biplot(icecream,"age","creme_glacee_chocolat")
biplot(icecream,"age","taille")
}
