% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_improper_NAs}
\alias{clean_improper_NAs}
\title{Replace values improperly provided as a given string indicating missing value with NA}
\usage{
clean_improper_NAs(
  x,
  possible_strings = c("N/A", "[Nn]on [Rr]enseigné", "[Mm]anquant", "[Uu]nknown",
    "[Nn]ot provided", "[Mm]issing")
)
}
\arguments{
\item{x}{the vector}

\item{possible_strings}{the strings provided in raw table which could possibly indicate a NA value}
}
\description{
Replace values improperly provided as a given string indicating missing value with NA
}
