% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{is_free_text}
\alias{is_free_text}
\title{Tests whether vector is free text or not}
\usage{
is_free_text(x, max_distinct = 20)
}
\arguments{
\item{x}{the vector}

\item{max_distinct}{the maximum number of distinct values to consider a column as free text. Defaults to 20.}
}
\description{
Tests whether vector is free text or not
}
