% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{is_mainly_missing}
\alias{is_mainly_missing}
\title{Replace values improperly provided as a given string indicating missing value with NA}
\usage{
is_mainly_missing(x, threshold_proportion = 0.3333)
}
\arguments{
\item{x}{the vector}

\item{threshold_proportion}{the proportion of missing values above which the vector is considered as mainly missing. Defaults to 0.33.}
}
\description{
Replace values improperly provided as a given string indicating missing value with NA
}
