% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significance.R
\name{pval_graph}
\alias{pval_graph}
\title{Show all two-by-two test results for a dataset}
\usage{
pval_graph(dataset, var_reorder = FALSE)
}
\arguments{
\item{dataset}{the dataset with all variables to test two-by-two}

\item{var_reorder}{whether to reorder variables so as to have the ones with the most significant two-by-two relationships to other variables on top. Defaults to FALSE.}
}
\value{
a plot with all relationships two
}
\description{
Show all two-by-two test results for a dataset
}
\examples{
# example code
data(icecream)
pval_graph(icecream \%>\% dplyr::select(-comment,-id))
}
