% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{Runs a random forest model}
\usage{
run_rf(data, response = colnames(data)[1], plot = FALSE, min_importance = 10)
}
\arguments{
\item{data}{the data (optimized for running a rf)}

\item{response}{the name of the response variable}

\item{plot}{whether or not to plot the importance of predictors. Defaults to FALSE}
}
\description{
Runs a random forest model
}
\examples{
# example code
data(icecream)
run_rf(icecream \%>\%
          dplyr::select(-id) \%>\%
          dplyr::mutate_all(optimize_for_rf),
          response="regular_eating",
          plot=TRUE)

run_rf(icecream \%>\%
          dplyr::select(-id) \%>\%
          dplyr::mutate_all(optimize_for_rf),
          response="taille",
          plot=TRUE)
}
