% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot}
\alias{biplot}
\title{Performs a plot crossing two variables, adapting to the kind of variable (qualitative or quantitative)}
\usage{
biplot(dataset, x, y)
}
\arguments{
\item{x}{the name of first variable}

\item{y}{the name of second variable}
}
\description{
Performs a plot crossing two variables, adapting to the kind of variable (qualitative or quantitative)
}
\examples{
# example code
data(icecream)
biplot(icecream,"creme_glacee_vanille","creme_glacee_chocolat")
biplot(icecream,"age","creme_glacee_chocolat")
biplot(icecream,"age","taille")
}
