% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{optimize_for_rf}
\alias{optimize_for_rf}
\title{Transforms a vector for ease of use in a random forest}
\usage{
optimize_for_rf(x, replace_NA_with = "Not provided")
}
\arguments{
\item{x}{the vector to optimize for running a random forest}

\item{replace_NA_with}{the character string used to replace NA as category in categorical responses. Defaults to "Not provided"}
}
\description{
Transforms a vector for ease of use in a random forest
}
