% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence_plot.R
\name{partial_dependence_plot}
\alias{partial_dependence_plot}
\title{Plots the partial dependence of a random forest model response to a predictor (based on the training dataset)}
\usage{
partial_dependence_plot(rf, predictor, probabilities = FALSE)
}
\arguments{
\item{rf}{the result of a run_rf() function}

\item{predictor}{the name (passed as a string) of the predictor considered}

\item{probabilities}{if TRUE, the probabilities of the (categorical) response variable are plotted instead of the predicted values. Defaults to FALSE.}
}
\value{
a partial dependence plot
}
\description{
Plots the partial dependence of a random forest model response to a predictor (based on the training dataset)
}
\examples{
# example code
data(icecream)
lemon_rf=run_rf(icecream \%>\%
          dplyr::select(-id) \%>\%
          dplyr::mutate_all(optimize_for_rf),
          response="sorbet_citron")
partial_dependence_plot(lemon_rf,predictor="age")
partial_dependence_plot(lemon_rf,predictor="creme_glacee_vanille")
}
