% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{Runs a random forest model}
\usage{
run_rf(data, response = colnames(data)[1], plot = FALSE, clean_name = FALSE)
}
\arguments{
\item{data}{the data (optimized for running a rf)}

\item{response}{the name of the response variable}
}
\description{
Runs a random forest model
}
\examples{
# example code
data(icecream)
run_rf(icecream \%>\%
          select(-id) \%>\%
          mutate_all(optimize_for_rf),
          response="sorbet_citron")
}
