% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rf.R
\name{run_rf}
\alias{run_rf}
\title{Tests whether vector is free text or not}
\usage{
run_rf(data, reponse = colnames(data)[1], plot = FALSE, clean_name = FALSE)
}
\arguments{
\item{x}{the vector to optimize for running a random forest}

\item{replace_NA_with}{the character string used to replace NA as category in categorical responses. Defaults to "Not provided"}
}
\description{
Tests whether vector is free text or not
}
\examples{
# example code
data(icecream)
run_rf(icecream,reponse="sorbet_citron")
}
